+++ apps/gesha/src/main.rs
@@ -1,10 +1,13 @@
 use clap::Parser;
 use gesha_core::gateway;
-use gesha_core::gateway::{Reader, Writer};
+use gesha_core::gateway::Reader;
 use gesha_core::targets::rust_type::Modules;
 use openapi_types::v3_0;
 use std::process::exit;
-use Subcommand::{Generate, GenerateSample};
+use Subcommand::{Generate, Test};
+
+mod tests;
+use tests::RunTestsArgs;
 
 fn main() {
     let args: Args = Args::parse();
@@ -12,7 +15,7 @@ fn main() {
 
     let result = match args.sub {
         Generate(x) => generate(x),
-        GenerateSample(x) => generate_sample(x),
+        Test(x) => tests::run_tests(x),
     };
     result.unwrap_or_else(|e| {
         println!("[failed] {:#?}", e);
@@ -30,7 +33,7 @@ struct Args {
 #[derive(clap::Subcommand, Debug)]
 enum Subcommand {
     Generate(GenerateArgs),
-    GenerateSample(GenerateSampleArgs),
+    Test(RunTestsArgs),
 }
 
 #[derive(clap::Args, Debug)]
@@ -39,38 +42,11 @@ struct GenerateArgs {
     schema: String,
 }
 
-#[derive(clap::Args, Debug)]
-struct GenerateSampleArgs {
-    #[clap(long)]
-    schema: String,
-
-    #[clap(long)]
-    output: String,
-}
-
 fn generate(args: GenerateArgs) -> gateway::Result<()> {
     println!("generate> {:?}", args);
 
     let reader = Reader::new::<v3_0::Document>();
     let rust_types: Modules = reader.open_rust_type(args.schema)?;
     println!("components: {:#?}", rust_types);
-    Ok(())
-}
-
-fn generate_sample(args: GenerateSampleArgs) -> gateway::Result<()> {
-    println!("generate_sample> {:?}", args);
-
-    let reader = Reader::new::<v3_0::ComponentsObject>();
-    let rust_types: Modules = reader.open_rust_type(args.schema)?;
-
-    println!("components: {:#?}", rust_types);
-
-    let writer = Writer {
-        path: args.output.into(),
-        preamble: None,
-    };
-    writer.print(rust_types)?;
-
-    println!("[done]");
     Ok(())
 }
